Switch to the new syntax and editing mechanism for datasets
Datasets are stored and edited separatedly, and are not part
of the manifest.

Also, switch back to raw JSON for the editor
