+++ views/thingpedia_device_create_or_edit.pug
@@ -7,9 +7,7 @@ block styles
 
 block scripts
   script(src='/javascripts/jsonlint.js')
-  script(src='/javascripts/codemirror.js')
-  script(src='/javascripts/jsoneditor.min.js')
-  script(src='/javascripts/thingpedia-device-create.js')
+  script(src='/javascripts/thingpedia-device-create-bundle.js')
 
 block content
   if create
@@ -50,7 +48,7 @@ block content
 
           div.form-group
             label(for='device-description-input').control-label= _("Device Description")
-            textarea(name='description',required=true).form-control#device-description #{device.description}
+            textarea(name='description',required=true).form-control#device-description= device.description
 
           div.form-group
             label(for='device-icon').control-label= _("Device Icon")
@@ -60,16 +58,13 @@ block content
             label(for='device-zipfile').control-label= _("Upload ZIP file with JS code")
             input(type='file',name='zipfile')#device-zipfile
 
-          div.form-group.hidden
+          div.form-group
             label(for='device-code').control-label Device Manifest
-            textarea(name='code',required=true).form-control#device-code #{device.code}
-            //span.help-block If your thing needs parameters for configurations (such as user names or
-              |  passwords), you must declare them in the manifest. You can also use the manifest
-              |  to declare actions and triggers for simple REST based services with a JSON API.
-              |  In any case, you should use the manifest to document how your device is used, that
-              |  is, what triggers and what actions are there!
+            textarea(name='code',required=true).form-control#device-code= device.code
 
-          div#json-manifest-placeholder
+          div.form-group
+            label(for='device-dataset').control-label= _("Default Dataset")
+            textarea(name='code',required=true).form-control#device-dataset= device.dataset
 
           if user.developer_status >= 2
             div.checkbox
